requirements.txt
pymongo
PySide6
reportlab
matplotlib
pandas
python-dotenv
requests